fixed some bugs, added query for funds needed
