+++ contracts/governance_admin/src/query/mod.rs
@@ -1,1 +1,2 @@
 pub mod query_refundable_funds;
+pub mod query_pool_creation_funds;
