+++ contracts/governance_admin/src/utils/constants.rs
@@ -1,0 +1,3 @@
+
+/// this address is derived using: https://gist.github.com/xlab/490d0e7937a8ccdbf805acb00f5dd9a1
+pub const GOV_MODULE_ADDRESS: &str = "persistence10d07y265gmmuvt4z0w9aw880jnsr700j5w4kch";
