+++ contracts/governance_admin/tests/utils/mod.rs
@@ -9,7 +9,7 @@ use dexter::vault::FeeInfo;
 
 use dexter::vault::{PauseInfo, PoolCreationFee, PoolType, PoolTypeConfig};
 
-use dexter_governance_admin::contract::GOV_MODULE_ADDRESS;
+use dexter_governance_admin::utils::constants::GOV_MODULE_ADDRESS;
 use persistence_std::types::cosmos::bank::v1beta1::QueryBalanceRequest;
 use persistence_std::types::cosmos::gov::v1::{
     MsgSubmitProposal, MsgVote, ProposalStatus, QueryProposalRequest, VoteOption,
